32 bits too small for books scoring

+++ hashcode2020_qualification_task/src/main/java/Book.java
@@ -1,8 +1,8 @@
 public class Book {
     final int id;
-    final int score;
+    final long score;
 
-    public Book(int id, int score) {
+    public Book(int id, long score) {
         this.id = id;
         this.score = score;
     }
